+++ packages/a11y/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @interunit/a11y
 
+## 0.0.13
+
+### Patch Changes
+
+- Emit type fix?
+- Updated dependencies
+  - @interunit/primitives@0.0.14
+  - @interunit/config@0.0.10
+
+## 0.0.12
+
+### Patch Changes
+
+- Attempting to fix type emit
+- Updated dependencies
+  - @interunit/primitives@0.0.13
+  - @interunit/config@0.0.9
+
 ## 0.0.11
 
 ### Patch Changes