+++ packages/combobox/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @interunit/combobox
 
+## 0.0.9
+
+### Patch Changes
+
+- Emit type fix?
+- Updated dependencies
+  - @interunit/primitives@0.0.14
+  - @interunit/config@0.0.10
+
+## 0.0.8
+
+### Patch Changes
+
+- Attempting to fix type emit
+- Updated dependencies
+  - @interunit/primitives@0.0.13
+  - @interunit/config@0.0.9
+
 ## 0.0.7
 
 ### Patch Changes