+++ packages/tabs/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @interunit/tabs
 
+## 0.0.7
+
+### Patch Changes
+
+- Emit type fix?
+- Updated dependencies
+  - @interunit/primitives@0.0.14
+  - @interunit/toolbox@0.0.9
+  - @interunit/config@0.0.10
+  - @interunit/a11y@0.0.13
+
+## 0.0.6
+
+### Patch Changes
+
+- Attempting to fix type emit
+- Updated dependencies
+  - @interunit/primitives@0.0.13
+  - @interunit/toolbox@0.0.8
+  - @interunit/config@0.0.9
+  - @interunit/a11y@0.0.12
+
 ## 0.0.5
 
 ### Patch Changes